+++ inst/examples/dungeonheroes/modules/realms/grey_mountains.R
@@ -4,7 +4,7 @@
     hide_map_navigation()
     game$activate_map(
       current_realm, player_name = "hero", x = 100, y = 100,
-      hidden_objects = c(mushroom_swamps_objects, "talk_bubble_text", "blacksmith")
+      hidden_objects = c(mushroom_swamps_objects, wild_forests_objects, "talk_bubble_text", "blacksmith")
     )
     enemy_status_text$set("enemies: none in grey mountains")
     set_combat_status("Explore the grey mountains.")