+++ inst/examples/dungeonheroes/modules/realms/mushroom_swamps.R
@@ -5,7 +5,7 @@
     game$activate_map(
       "mushroom_swamps", player_name = "hero", x = 100, y = 100,
       visible_objects = mushroom_swamps_objects,
-      hidden_objects = "blacksmith"
+      hidden_objects = c(wild_forests_objects, "blacksmith")
     )
     update_enemy_status()
     set_combat_status("Back in the mushroom swamps.")