+++ inst/examples/dungeonheroes/modules/realms/wild_forests.R
@@ -4,6 +4,7 @@
     hide_map_navigation()
     game$activate_map(
       current_realm, player_name = "hero", x = 100, y = 100,
+      visible_objects = wild_forests_objects,
       hidden_objects = c(mushroom_swamps_objects, "talk_bubble_text", "blacksmith")
     )
     enemy_status_text$set("enemies: none in wild forests")